back-end/requirements.txt
asgiref==3.7.2
attrs==21.2.0
blinker==1.4
certifi==2023.7.22
charset-normalizer==3.2.0
command-not-found==0.3
cryptography==3.4.8
dbus-python==1.2.18
distlib==0.3.6
distro==1.7.0
distro-info===1.1build1
Django==4.2.4
django-cors-headers==4.2.0
djangorestframework==3.14.0
exceptiongroup==1.1.1
filelock==3.12.2
gyp==0.1
httplib2==0.20.2
idna==3.4
importlib-metadata==4.6.4
iniconfig==2.0.0
jeepney==0.7.1
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
more-itertools==8.10.0
netifaces==0.11.0
oauthlib==3.2.0
packaging==23.1
platformdirs==3.8.1
pluggy==1.0.0
psutil==5.9.5
psycopg==3.1.10
psycopg-binary==3.1.10
py==1.10.0
pygame==2.5.0
Pygments==2.11.2
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==2.4.7
pytest==7.3.2
pytest-mock==3.10.0
pytest-print==0.3.1
python-apt==2.4.0+ubuntu1
python-dotenv==1.0.0
pytz==2023.3
PyYAML==5.4.1
requests==2.31.0
SecretStorage==3.3.1
six==1.16.0
sqlparse==0.4.4
systemd-python==234
toml==0.10.2
tomli==2.0.1
trash-cli==0.23.2.13.2
typing==3.7.4.3
typing_extensions==4.7.1
ubuntu-advantage-tools==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==2.0.4
virtualenv==20.23.1
wadllib==1.3.6
zipp==1.0.0
